Hauptseminar-AutoGluon/ShamilNabiyev/requirements.txt
argcomplete==1.12.3
atomicwrites==1.4.0
attrs==21.2.0
autogluon.core==0.3.1
autogluon.features==0.3.1
autogluon.tabular==0.3.1
autograd==1.3
backcall==0.2.0
bcrypt==3.2.0
blis==0.7.5
boto3==1.15.18
botocore==1.18.18
catalogue==2.0.6
catboost==0.25.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
ConfigSpace==0.4.19
cryptography==35.0.0
cycler==0.10.0
cymem==2.0.6
Cython==0.29.24
dask==2021.9.1
debugpy==1.5.1
decorator==5.1.0
dill==0.3.4
distributed==2021.9.1
entrypoints==0.3
fastai==2.5.2
fastcore==1.3.26
fastdownload==0.0.5
fastprogress==1.0.0
fsspec==2021.10.1
future==0.18.2
graphviz==0.8.4
HeapDict==1.0.1
idna==2.6
importlib-metadata==4.8.1
iniconfig==1.1.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
kiwisolver==1.3.2
lightgbm==3.3.0
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
msgpack==1.0.2
murmurhash==1.0.6
mxnet==1.7.0.post2
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.21.3
packaging==21.0
pandas==1.3.4
paramiko==2.8.0
parso==0.8.2
partd==1.2.0
pathy==0.6.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
pluggy==1.0.0
preshed==3.0.6
prompt-toolkit==3.0.21
psutil==5.8.0
py==1.10.0
pycowsay==0.0.0.1
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
PyYAML==6.0
pyzmq==22.3.0
requests==2.18.4
s3transfer==0.3.7
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
smart-open==5.2.1
sortedcontainers==2.4.0
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.2
tblib==1.7.0
tenacity==8.0.1
thinc==8.0.11
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
torch==1.9.1
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.22
wasabi==0.8.2
wcwidth==0.2.5
xgboost==1.4.2
zict==2.0.0
zipp==3.6.0
